+++ app/models/user.ts
@@ -15,6 +15,9 @@ export default class User extends compose(BaseModel, AuthFinder) {
   declare id: number
 
   @column()
+  declare username: string
+
+  @column()
   declare email: string
 
   @column({ serializeAs: null })
