--- app/validators/auth.ts
@@ -1,14 +1,0 @@
-import vine from '@vinejs/vine'
-
-export const registerValidator = vine.compile(
-  vine.object({
-    email: vine
-      .string()
-      .email()
-      .unique(async (query, field) => {
-        const user = await query.from('users').where('email', field).first()
-        return !user
-      }),
-    password: vine.string().minLength(3).maxLength(64),
-  })
-)
