+++ app/validators/auth/register_validator.ts
@@ -1,0 +1,28 @@
+import vine from '@vinejs/vine'
+
+export const registerValidator = vine.compile(
+  vine.object({
+    username: vine
+      .string()
+      .minLength(3)
+      .maxLength(32)
+      .regex(/^\w+$/)
+      .unique(async (query, field) => {
+        const user = await query.from('users').where('username', field).first()
+        return !user
+      }),
+    email: vine
+      .string()
+      .email()
+      .unique(async (query, field) => {
+        const user = await query.from('users').where('email', field).first()
+        return !user
+      }),
+    password: vine
+      .string()
+      .minLength(3)
+      .maxLength(32)
+      .regex(/^(?=.*[A-Z])(?=.*[a-z])(?=.*\d)(?=.*[^A-Za-z\d]).+$/)
+      .confirmed(),
+  })
+)
